requirements.txt
python-dotenv
discord.py
tabulate
python-dateutil
aiohttp
apscheduler
SQLAlchemy